requirements.txt
requests
beautifulsoup4
pandas
tqdm
openpyxl
urllib3
streamlit
numpy
matplotlib
seaborn
plotly
wordcloud
textblob
spacy
langdetect
newspaper3k